+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - **add** : Supports new minimal version : `16.0.0`
+- **add** : Added implementation of GetAvailablePaymentProduct Endpoint API
 - **fix** : Fixed v3 format when requesting transaction information
 
 ## 1.5.0